Move atom tests into folder, create molecule tests for diagnosis redacted
